refactor: model에서 역할에 따라 패키지 분리
