+++ src/main/java/racingcar/model/score/ScoreBoard.java
@@ -1,4 +1,4 @@
-package racingcar.model;
+package racingcar.model.score;
 
 import java.util.List;
 import java.util.stream.Collectors;
