+++ src/test/java/racingcar/model/car/CarTest.java
@@ -1,4 +1,4 @@
-package racingcar.model;
+package racingcar.model.car;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
