+++ src/test/java/racingcar/model/car/CarsTest.java
@@ -1,7 +1,10 @@
-package racingcar.model;
+package racingcar.model.car;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.model.car.Car;
+import racingcar.model.car.CarDto;
+import racingcar.model.car.Cars;
 
 import java.util.List;
